Adding permissions to trustee user
